Refactoring styled Components
