+++ src/components/Buttons/ButtonCancel.js
@@ -1,0 +1,11 @@
+import React from 'react';
+import { ButtonCancelStyle } from './ButtonGlobalStyle';
+
+const ButtonCancel = ({ onCancelDelete }) => {
+    return (
+        <ButtonCancelStyle onClick = { onCancelDelete }>Cancel</ButtonCancelStyle>
+    )
+}
+
+export default ButtonCancel;
+
