+++ src/components/Buttons/ButtonDelete.js
@@ -1,0 +1,10 @@
+
+import { ButtonDeleteStyle } from './ButtonGlobalStyle';
+
+const ButtonDelete = ({ deleteInvoice }) => {
+    return (
+        <ButtonDeleteStyle onClick = {deleteInvoice} >Delete</ButtonDeleteStyle>
+    )
+}
+
+export default ButtonDelete;
