+++ src/components/Buttons/FormButtons.js
@@ -1,13 +1,12 @@
 import React from 'react';
-import styled  from 'styled-components';
-
+import { ButtonsContainer, FirstButtons, ButtonCancelStyle, ButtonSaveStyle  } from './ButtonsGlobalStyle';
 const FormButtons = ({ discardForm, saveAsDraft, submitForm}) => {
 
     return(
         <ButtonsContainer>
             <FirstButtons>
-                <button onClick = {discardForm}>Discard</button>
-                <button onClick = { saveAsDraft } >Save as Draft</button>
+                <ButtonCancelStyle onClick = {discardForm}>Discard</ButtonCancelStyle>
+                <ButtonSaveStyle onClick = { saveAsDraft } >Save as Draft</ButtonSaveStyle>
             </FirstButtons>
                 <button onClick = {submitForm} >Save & Send</button>
         </ButtonsContainer>  
@@ -16,44 +15,5 @@ const FormButtons = ({ discardForm, saveAsDraft, submitForm}) => {
 
 export default FormButtons;
 
-const ButtonsContainer = styled.div` 
-    width: 100%;
-    display: flex;
-    justify-content: space-between;
-    align-items: center; 
-    padding:0 48px;
-    @media screen and (max-width: 600px){
-        justify-content: space-between;
-        padding: 0 16px;
-        button{
-            padding: 17px 16px;
-        }
-        
-    }
-`
-const FirstButtons = styled.section`
-    width: 72%;
-    display:flex;
-    justify-content: space-between;
-    button{
-        &:nth-of-type(1){
-            background-color: #F9FAFE;
-            color: #7E88C3;
-            &:hover{
-                background-color: #DFE3FA;
-            }
-        }
-        &:nth-child(2){
-            background-color: #373B53;
-            color: #888EB0;
-            margin-right: 6px;
-            &:hover{
-                background-color: #0C0E16;
-            }
-        }        
-    }
-    @media screen and (max-width: 600px) {
-        width: 65%;
-    }
-`
+
