+++ src/components/Empty.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import styled  from 'styled-components';
+import {EmptySection}  from './EmptyStyle';
 import illustration from '../assets/illustration-empty.svg';
 
 const Empty = () => {
@@ -15,28 +15,3 @@ const Empty = () => {
 }
 
 export default Empty;
-
-const EmptySection = styled.section` 
-    margin: 140px auto;
-    width: 242px;
-    text-align: center;
-    span{
-        font-weight: 700;
-    }
-    h2{
-         color: ${props => props.theme.titleColor};// #0C0E16;//#FFFFFF;
-         margin-bottom: 24px;
-    }
-    p{
-        color: ${props => props.theme.paragraphColor};// #888EB0;//#DFE3FA
-    }
-    @media screen and (max-width:600px){
-        margin: 102px auto;
-        img{
-            width: 193px;
-        }
-        span span{
-                display: none;            
-        }
-    }
- `
