+++ src/components/Forms/SelectForm.js
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import styled  from 'styled-components';
+import { FormStyling }  from './SelectFormStyle';
 import arrowDown from '../../assets/icon-arrow-down.svg';
 import InputRadio from './InputRadio'; 
 
@@ -34,33 +34,3 @@ const SelectForm = ({onStatusSelected, theme}) => {
 
 export default SelectForm;
 
-const FormStyling = styled.form` 
-    margin-right: 40px;    
-    position: relative;
-    label{
-        font-weight: 700;    
-        color: ${props => props.theme.titleColor};        
-        cursor: pointer; 
-    }
-    img{
-        margin-left: 16px;
-    }
-    fieldset{
-        position: absolute;
-        top: 24px;
-        left: -32px;
-        width: 192px;
-        height: 128px;
-        padding: 8px;
-        border: 0;
-        box-shadow:${props => props.theme.shadowField};
-        border-radius: 8px;
-        background : ${props => props.theme.backgroundField};
-    }
-    @media screen and (max-width:600px){
-        margin-right: 12px;
-        span:first-child{
-            display: none;
-        }
-    }
- `
