+++ src/components/GoBack.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import styled  from 'styled-components';
+import { LinkStyle }  from './GoBackStyle';
 import { Link } from 'react-router-dom';
 import arrowLeft from '../assets/icon-arrow-left.svg';
 
@@ -16,13 +16,3 @@ const GoBack = () => {
 
 export default GoBack;
 
-const LinkStyle = styled.p`     
-        font-weight: 700;
-        color: ${props => props.theme.titleColor};
-        img{
-            margin-right: 24px;
-        }
-        &:hover{
-            color: ${props => props.theme.paragraphColor};
-        }    
-`
