+++ src/components/Headers/Header.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import styled  from 'styled-components';
+import { HeaderContainer, FormContainer }  from './HeaderStyle';
 import SelectForm from '../Forms/SelectForm';
 import AddButton from '../Buttons/AddButton';
 
@@ -22,26 +22,5 @@ const Header = ({num, onStatusSelected, theme, handleAddForm}) => {
 
 export default Header;
 
-const HeaderContainer = styled.header` 
-    width: 100%;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    h1{
-         color: ${props => props.theme.titleColor};
-    }
-    p{
-        color: ${props => props.theme.paragraphColor};
-        @media screen and (max-width: 600px) {
-            span{
-                display: none;
-            }
-        }
-    }    
- `
- const FormContainer = styled.div`  
-    display: flex;
-    align-items: center;
- `
-  
+
