+++ src/components/Headers/InvoiceHeader/HeaderStatus.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import styled  from 'styled-components';
+import { StatusContainer, StatusStyle }  from './HeaderStatusStyle';
 import Status from '../../Invoices/Status';
 
 const HeaderStatus = ({status}) => {
@@ -16,20 +16,3 @@ const HeaderStatus = ({status}) => {
 
 export default HeaderStatus;
 
-const StatusContainer = styled.div`
-width: 159px;
-display: flex;
-justify-content: space-between;
-align-items: center;
-color: "#858BB2";
-@media screen and (max-width: 600px){
-    width: 100%;
-}    
-`
-const StatusStyle = styled.div`
-    margin-left: 5px;
-    @media screen and (max-width: 600px){
-        margin-top: 24px;
-        margin-left: 0;
-    }
- `
