+++ src/components/Main.js
@@ -1,5 +1,5 @@
 import React, { useState }from 'react';
-import styled  from 'styled-components';
+import { InvoicesListContainer, MainContainer } from './MainStyle';
 import Header from './Headers/Header';
 import Empty from './Empty';
 import InvoiceListItem from './Invoices/InvoiceListItem';
@@ -38,24 +38,4 @@ const onStatusSelected = (status) => {
 export default Main;
 
 
-const InvoicesListContainer = styled.ul`
-    width: 100%;
-    margin-top: 65px;       
-`
-
-const MainContainer = styled.main`
-    width: 51%;
-    margin: 72px auto;
-    position : relative;
-    @media screen and (max-width: 1200px) {
-        width: 70%;
-    }
-    @media screen and (max-width: 1000px) {
-        width: 88%;
-        margin: 56px auto;
-    }
-    @media screen and (max-width: 600px) {
-       margin: 32px auto;
-    }
-`
