+++ src/pages/Invoice.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import { useLocation } from 'react-router-dom';
-import styled  from 'styled-components';
+import { Container, InvoiceFooter } from './InvoiceStyle'
 import Aside from '../components/Aside';
 import InvoiceContainer from './../components/Invoices/InvoiceContainer';
-import Buttons from '../components/Headers/InvoiceHeader/Buttons';
+import Buttons from '../components/Buttons/Buttons';
 
 
 const Invoice = ({theme, setTheme, deleteInvoice, invoices, onUpdateForm, onClickDelete, onClickPaid,}) => {
@@ -32,23 +32,3 @@ console.log(invoices)
 
 export default Invoice;
 
-const Container = styled.div` 
-    width: 100%;
-    min-height: 100vh;
-    background-color: ${props=>props.theme.background};// #f2f2f2;//141625
-    display: flex;
-    transition: all 0.5s ease;
-    @media screen and (max-width: 1000px) {
-        flex-direction: column;
-    }
-`
-const InvoiceFooter = styled.footer`
-    display: none;
-    margin-top: 20px;
-    @media screen and (max-width: 600px){
-        width: 100%;
-        background: ${props => props.theme.backgroundInvoice};
-        padding: 24px 8%;            
-        display: inline-block;
-    }
-`
