+++ src/pages/ModalEdit.js
@@ -1,75 +1,17 @@
 import React, {useState, useRef, useEffect } from 'react';
-import styled  from 'styled-components';
 import bin  from '../assets/icon-delete.svg';
 import EditButtons  from '../components/Buttons/EditButtons';
 import GoBack  from '../components/GoBack';
 import FormErrors from '../components/Errors/FormErrors';
+import { Invoice } from '../util/Invoice';
+import { Item } from '../util/Item';
+import { formValidation } from '../util/formValidation';
+import { createId } from '../util/createId';
+import  {ModalContainer, LinkContainer , FormContainer, FlexWrapper , ItemField , ButtonAddItem,  
+    GradientDiv , ErrorsStyling } from './ModalStyle';
 
 const ModalEdit = ({invoice,  onUpdateForm , handleGoBack}) => {
-    
-    class Address  {
-        constructor(street = "", city = "", postCode = "", country = ""){
-            this.street = street;
-            this.city =  city;
-            this.postCode =  postCode;
-            this.country = country;
-        }
-    }
 
-    class Item {
-        constructor(name = "", quantity = "0", price = "0"){
-            this.name = name;
-            this.quantity = quantity;
-            this.price =  price;
-            this.total = this.calculateTotal();              
-        }
-        calculateTotal = () => (this.price*1) * (this.quantity*1);
-    }
-
-    class Invoice{
-        constructor(createdAt = new Date(), status,  description = '', paymentTerms = "1", 
-                    clientName = "", clientEmail = "", senderStreet = "", senderCity ="", 
-                    senderPostCode = "", senderCountry = "", clientStreet = "", clientCity = "", 
-                    clientPostCode = "", clientCountry = "", items = []){
-            
-            this.createdAt = createdAt;
-            this.paymentTerms = paymentTerms;
-            this.paymentDue = this.calculatePaymentDue();
-            this.description = description;            
-            this.clientName = clientName;
-            this.clientEmail = clientEmail;
-            this.status = status;
-            this.senderAddress  = new Address(senderStreet, senderCity, senderPostCode, senderCountry);
-            this.clientAddress = new Address(clientStreet, clientCity, clientPostCode, clientCountry);
-            this.items = items;
-            this.total = 0;
-        }
-
-        calculatePaymentDue = () =>{
-            if(this.createdAt && this.paymentTerms){
-                let days = this.paymentTerms*(24*60*60*1000);
-                let invoiceDate = Date.parse(this.createdAt);
-                return new Date(invoiceDate + days);
-            }else{
-                return new Date()
-            }            
-        } 
-
-        
-        addItem = (item) => {
-            this.items.push(item)
-        }
-         calculateTotal = () => {
-             let sum = 0;
-             console.log(this.items)
-             for(let i = 0; i < this.items.length; i++){
-                sum += this.items[i].total*1
-             }
-             this.total =  sum;
-         }
-         
-    }
-    
     const createdAt = useRef('');
     const description = useRef('');
     const paymentTerms = useRef('');
@@ -83,9 +25,6 @@ const ModalEdit = ({invoice,  onUpdateForm , handleGoBack}) => {
     const clientCity = useRef('');
     const clientCountry = useRef('');
     const clientPostCode = useRef('');
-    const itemName = useRef('');
-    const itemQuantity = useRef('');
-    const itemPrice = useRef('');
     
     const initialInvoice = JSON.parse(JSON.stringify(invoice))
     const [ invoiceEdit, setInvoiceEdit] = useState(new Invoice(initialInvoice.createdAt, 
@@ -101,45 +40,10 @@ const ModalEdit = ({invoice,  onUpdateForm , handleGoBack}) => {
         
         
     const [formErrors, setFormErrors] = useState([]);
-
-    const [isFormValid, setIsFormValid] = useState(false);
     const [isSubmitted, setIsSubmitted] = useState(false);
 
 
-    const formValidation = (formValues, fieldNames, fieldRefs) =>{
-        let fieldValidationErrors = [];
-        for(let i=0; i < 13; i++){         
-            switch(fieldNames[i]) {
-              case 'clientName':
-                //let nameValid = (formValues[i].match(/^[a-zA-ZŠšĐđŽžČčĆć]+$/) 
-                let nameValid = (formValues[i].match(/^[a-z ,.'-]+$/i) 
-                && formValues[i] !=="");
-                if(!nameValid) {
-                    fieldValidationErrors.push("Client name is not valid!");
-                    fieldRefs[i].current.parentElement.className = "error";
-                    }
-                break;
-               case 'clientEmail':
-                let emailValid = (formValues[i].match(/^([\w.%+-]+)@([\w-]+\.)+([\w]{2,})$/i) 
-                && formValues[i] !=="");
-                if(!emailValid) {
-                        fieldValidationErrors.push("Client email is invalid!");
-                        fieldRefs[i].current.parentElement.className = "error";
-                    }
-                break;
-              default:
-                  let fieldValid = (formValues[i] !== '');
-                  if(!fieldValid) {
-                      fieldValidationErrors.push(fieldNames[i] + " cannot be empty!")
-                      fieldRefs[i].current.parentElement.className = "error";
-                  }   
-                break;
-            }
-         }
-       return fieldValidationErrors;       
-    }
-         
-
+    
         const handleOnChange = (index, event) => {
             const values = [...itemFields];      
             values[index][event.target.name] = event.target.value;
@@ -153,16 +57,12 @@ const ModalEdit = ({invoice,  onUpdateForm , handleGoBack}) => {
         }
 
         const removeItemField = (id) => {
-            console.log(id)
             const values = [...itemFields];
             values.splice(id, 1);
             setItemFields(values);
         }
 
-        const  createId = () => "ABCDEFGHIJKLMNOPQRSTUVWXYZ"[ Math.floor(Math.random() *25)] 
-            + "ABCDEFGHIJKLMNOPQRSTUVWXYZ"[ Math.floor(Math.random() *25)] 
-                + Math.floor(Math.random() * 9999).toString().padStart(4,0);          
-
+    
         const createInvoice = () => {
             const createdAtValue = createdAt.current.value;
             const descriptionValue = description.current.value;
@@ -205,7 +105,8 @@ const ModalEdit = ({invoice,  onUpdateForm , handleGoBack}) => {
     const onFormSubmit = (event) => {
         event.preventDefault();
         const [addedInvoice, errorsList] = createInvoice();
-        setInvoiceEdit(addedInvoice)
+        addedInvoice.status = "pending";
+        setInvoiceEdit(addedInvoice);
         setFormErrors(errorsList);
         setIsSubmitted(true);        
     }
@@ -213,24 +114,14 @@ const ModalEdit = ({invoice,  onUpdateForm , handleGoBack}) => {
     useEffect(() => {
         if (isSubmitted) {
           if (formErrors.length === 0) {
-            setIsFormValid(true);
             onUpdateForm(invoiceEdit);
             handleGoBack();
-          } else {
-            setIsFormValid(false);
-
-          }
+          } 
           setIsSubmitted(false);
         }
       }, [isSubmitted, formErrors, handleGoBack, invoiceEdit, onUpdateForm]);
     
-      
-         
-
-    const handleSubmit = (event) => {
-        event.preventDefault();
-    }
-
+     
     const onFormCancel = () => {
         setInvoiceEdit(initialInvoice);
         handleGoBack();
@@ -410,215 +301,3 @@ const ModalEdit = ({invoice,  onUpdateForm , handleGoBack}) => {
 
 export default ModalEdit;
 
-const ModalContainer = styled.section` 
-    width: 100%;
-    height: 100%;
-    position: absolute;
-    left: 103px;
-    display: flex;
-    background: rgba(0, 0, 0, 0.5);
-    overflow: scroll;
-    @media screen and (max-width: 1000px) {
-        top: 80px;
-        left: 0;
-    }
-    @media screen and (max-width: 600px) {
-        top: 72px;
-    }
-`
-const LinkContainer = styled.div` 
-    display: none;
-    margin: 24px;
-    @media screen and (max-width:600px){
-        display: block;
-    }
-`
-const FormContainer = styled.form`
-    width: 616px;
-    height: 100%;
-    background-color: ${props => props.theme.backgroundAddInvoice};
-    overflow: scroll;
-    border-radius: 0 28px 28px 0;
-    padding: 8px;
-    
-    h1{
-        font-size: 24px;
-        line-height: 32px;
-        letter-spacing: -0.5px;
-        margin: 48px;
-        color: ${props => props.theme.titleColor};
-    }
-    h2{
-        font-size: 18px;
-        line-height: 32px;
-        letter-spacing: -0.38;
-        color: #777F98;       
-    }
-    fieldset{
-       margin: 48px;
-        border: none; 
-    }
-    label{
-        margin-top: 24px;
-        display: block;
-        color: #7E88C3;
-    }
-    legend{
-        color:#7C5DFA;
-        font-weight: 700;        
-    }
-    label{
-        color: ${props => props.theme.paragraphInvoice};
-    }
-    input, select{
-        color: ${props => props.theme.titleColor} !important;
-        border: ${props => props.theme.inputBorder};
-        box-shadow : ${props => props.theme.inputBoxShadow} ;
-    }
-    span{
-        float: right;
-        font-size: 10px;
-        line-height: 15px;
-        letter-spacing: -0.25px;
-        display: none;
-    }
-    .error{
-        color: #EC5757; 
-        span{
-            display: block;
-        }
-        input{            
-            border: 1px solid #EC5757;
-        }
-    }
-    input:-webkit-autofill{
-        -webkit-text-fill-color: ${props => props.theme.titleColor} !important;
-    }
-    input:hover, select:hover{
-       border: 1px solid #7C5DFA; 
-    }    
-    @media screen and (max-width:600px){
-        width: 100%;
-        h1{
-            margin: 24px 16px;
-        }
-        fieldset{
-            margin: 24px 16px;
-        }
-    }
- `
- const FlexWrapper = styled.div`
-    display: flex;
-    justify-content: space-between;
-    flex-wrap: wrap;
-    label{
-        width: 30%;        
-    }
-    label[for = "invoiceDate"]{
-        width: 48%;
-    }
-    label[for = "paymentTerms"]{
-        width: 48%;
-    }
-     
-    img{
-        object-fit: contain;
-        margin-top: 50px;
-        cursor: pointer;
-    }
-    @media screen and (max-width:600px){
-        label{
-            width: 152px;
-            &:nth-child(3){
-                width: 100%;
-            }
-        }
-        label[for = "invoiceDate"]{
-            width: 100%;
-        }
-        label[for = "paymentTerms"]{
-            width: 100%;
-        }
-    
-    }
-  `
-    const ItemField = styled.label`
-        &:nth-child(1){
-        width: 42%;
-    } 
-    &:nth-child(2){
-        width: 9%;
-    } 
-    &:nth-child(3){
-        width: 20%;
-    } 
-    &:nth-child(4){
-        width: 9%;
-        input{
-            border: transparent;
-            padding-left: 0;
-            color: ${props => props.theme.paragraphInvoice} !important;
-        }
-    } 
-    &:nth-child(5){
-        width: 15px;
-    }
-    @media screen and (max-width:600px){        
-        &:nth-child(1){
-            width: 100%;
-        }
-        &:nth-child(2){
-            width: 21%;
-        } 
-        &:nth-child(3){
-            width: 33%;
-        }
-        &:nth-child(4){
-            width: 20%;
-        }
-    }
-    `
-
-  const ButtonAddItem = styled.button`   
-        width: 100%;
-        background-color:${props => props.theme.backgroundItemsInvoice};
-        color: ${props => props.theme.paragraphInvoice};
-        margin: 16px 0 47px 0;
-        &:hover{
-            background-color: #DFE3FA ;
-        }
-        @media screen and (max-width:600px){
-            margin: 48px 0 0 0;
-        }
-  `
-  const GradientDiv = styled.div`
-        width: 100%;
-        height: 155px;
-        div{
-            &:first-child{
-                height:64px;
-                margin-bottom: 21px;
-                background-image: linear-gradient(180deg, rgba(0, 0, 0, 0.0001) 0%, rgba(0, 0, 0, 0.1) 100%);
-                display: none;
-            }
-        }
-        
-        @media screen and (max-width:600px){
-            div{
-                &:first-child{
-                    display: block;
-                }
-            }            
-        }
-   `
-   const ErrorsStyling = styled.div`
-   margin: 48px -16px 48px 44px;
-       p{
-            color:#EC5757;
-            font-size: 10px;
-            line-height: 15px;
-            letter-spacing: -0.21px;
-            font-weight: 500;
-       }
-   `
-
