+++ src/util/Address.js
@@ -1,0 +1,9 @@
+export class Address  {
+    
+    constructor(street = "", city = "", postCode = "", country = ""){
+        this.street = street;
+        this.city =  city;
+        this.postCode =  postCode;
+        this.country = country;
+    }
+}
